pattern-matching: Use labeled, non-loop, blocks for easier hygiene.
We've been using `continue` to break out of a non-match, like this:

  do {
    // ...
    if (!/* match */) continue;
    // ...
    /* body */
  } while (0);

This can work fine, but would require adding some complexity because
any `continue` or `break` statement inside the body, if it doesn't
refer to an explicit label, requires rewriting to avoid having the
new `do`-loop capture it.

If instead we use `break LABEL`, then we can have the outer block be
*not a loop*:

  _label: {
    // ...
    if (!/* match */) break _label;
    // ...
    /* body */
  }

Now an unlabelled `break` or `continue` is unaffected, so the problem
is reduced to generating a unique label to use -- which is a general
problem we have general infrastructure for, in `scope.generateUid`.
So do that.